requirements.txt
pandas
sentence_transformers
faiss-cpu
pickle
dotenv
google.generativeai
flask
flask-cors